+++ libs/net/dal/Services/ReportInstanceService.cs
@@ -64,7 +64,7 @@ public class ReportInstanceService : BaseService<ReportInstance, long>, IReportI
     /// <param name="reportId"></param>
     /// <param name="date"></param>
     /// <returns></returns>
-    public ReportInstance? FindInstanceForReportIdAndDate(int reportId, DateTime date)
+    public ReportInstance? FindInstanceForReportIdAndDate(long reportId, DateTime date)
     {
         DateTime morning = new DateTime(date.Year, date.Month, date.Day, 0, 0, 0);
         DateTime midnight = morning.AddHours(24);
